rename multiplexer to container
