+++ src/components/Timeline/Message/MessageActions.tsx
@@ -25,7 +25,7 @@ import type { SimpleNote as TypeSimpleNote } from '../../../schemas/simpleNote'
 import { useRef, useState } from 'react'
 import Collapse from '@mui/material/Collapse'
 import Fade from '@mui/material/Fade'
-import { useMessageService } from '../Multiplexer'
+import { useMessageService } from '../MessageContainer'
 import ContentPasteIcon from '@mui/icons-material/ContentPaste'
 import ManageSearchIcon from '@mui/icons-material/ManageSearch'
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
